minor renames and version bumps
