+++ package.json
@@ -1,7 +1,7 @@
 {
 	"name": "godmode",
 	"productName": "GodMode",
-	"version": "1.0.0-beta.8",
+	"version": "1.0.0-beta.9",
 	"author": "smol-ai",
 	"description": "Simultaneously chat with ChatGPT, Claude, Bing, Bard, Llama2, et al in a full chat browser, NO API",
 	"main": "src/main/main.ts",
@@ -35,22 +35,22 @@
 	"dependencies": {
 		"@headlessui/react": "^1.7.17",
 		"@heroicons/react": "^2.0.18",
-		"@radix-ui/react-dialog": "^1.0.4",
+		"@radix-ui/react-dialog": "^1.0.5",
 		"@radix-ui/react-icons": "^1.3.0",
 		"@radix-ui/react-slot": "^1.0.2",
-		"@radix-ui/react-tooltip": "^1.0.6",
-		"chalk": "^4.1.2",
+		"@radix-ui/react-tooltip": "^1.0.7",
+		"chalk": "^5.3.0",
 		"class-variance-authority": "^0.7.0",
 		"clsx": "^2.0.0",
 		"electron-context-menu": "^3.6.1",
 		"electron-debug": "^3.2.0",
 		"electron-log": "^4.4.8",
 		"electron-store": "^8.1.0",
-		"electron-updater": "^6.1.1",
+		"electron-updater": "^6.1.4",
 		"react": "^18.2.0",
 		"react-beautiful-dnd": "^13.1.1",
 		"react-dom": "^18.2.0",
-		"react-router-dom": "^6.15.0",
+		"react-router-dom": "^6.16.0",
 		"react-split": "^2.0.14",
 		"tailwind-merge": "^1.14.0",
 		"vex-js": "^4.1.0"
@@ -65,38 +65,38 @@
 		"@pmmmwh/react-refresh-webpack-plugin": "^0.5.11",
 		"@svgr/webpack": "^8.1.0",
 		"@teamsupercell/typings-for-css-modules-loader": "^2.5.2",
-		"@testing-library/jest-dom": "^6.0.1",
+		"@testing-library/jest-dom": "^6.1.3",
 		"@testing-library/react": "^14.0.0",
-		"@types/jest": "^29.5.3",
-		"@types/node": "20.5.2",
-		"@types/react": "^18.2.20",
-		"@types/react-beautiful-dnd": "^13.1.4",
-		"@types/react-dom": "^18.2.7",
-		"@types/react-test-renderer": "^18.0.0",
+		"@types/jest": "^29.5.5",
+		"@types/node": "20.8.2",
+		"@types/react": "^18.2.25",
+		"@types/react-beautiful-dnd": "^13.1.5",
+		"@types/react-dom": "^18.2.10",
+		"@types/react-test-renderer": "^18.0.3",
 		"@types/terser-webpack-plugin": "^5.0.4",
-		"@types/webpack-bundle-analyzer": "^4.6.0",
-		"@typescript-eslint/eslint-plugin": "^6.4.1",
-		"@typescript-eslint/parser": "^6.4.1",
-		"autoprefixer": "^10.4.15",
+		"@types/webpack-bundle-analyzer": "^4.6.1",
+		"@typescript-eslint/eslint-plugin": "^6.7.4",
+		"@typescript-eslint/parser": "^6.7.4",
+		"autoprefixer": "^10.4.16",
 		"browserslist-config-erb": "^0.0.3",
 		"concurrently": "^8.2.1",
-		"core-js": "^3.32.1",
+		"core-js": "^3.33.0",
 		"cross-env": "^7.0.3",
 		"css-loader": "^6.8.1",
 		"css-minimizer-webpack-plugin": "^5.0.1",
 		"detect-port": "^1.5.1",
-		"electron": "^26.0.0",
-		"electron-builder": "^24.6.3",
+		"electron": "^26.3.0",
+		"electron-builder": "^24.6.4",
 		"electron-devtools-installer": "^3.2.0",
 		"electronmon": "^2.0.2",
-		"eslint": "^8.47.0",
+		"eslint": "^8.50.0",
 		"eslint-config-airbnb-base": "^15.0.0",
-		"eslint-config-erb": "^4.0.6",
-		"eslint-import-resolver-typescript": "^3.6.0",
+		"eslint-config-erb": "^4.1.0",
+		"eslint-import-resolver-typescript": "^3.6.1",
 		"eslint-import-resolver-webpack": "^0.13.7",
-		"eslint-plugin-compat": "^4.1.4",
+		"eslint-plugin-compat": "^4.2.0",
 		"eslint-plugin-import": "^2.28.1",
-		"eslint-plugin-jest": "^27.2.3",
+		"eslint-plugin-jest": "^27.4.2",
 		"eslint-plugin-jsx-a11y": "^6.7.1",
 		"eslint-plugin-promise": "^6.1.1",
 		"eslint-plugin-react": "^7.33.2",
@@ -104,16 +104,16 @@
 		"file-loader": "^6.2.0",
 		"html-webpack-plugin": "^5.5.3",
 		"identity-obj-proxy": "^3.0.0",
-		"jest": "^29.6.3",
-		"jest-environment-jsdom": "^29.6.3",
+		"jest": "^29.7.0",
+		"jest-environment-jsdom": "^29.7.0",
 		"mini-css-extract-plugin": "^2.7.6",
-		"postcss": "^8.4.28",
+		"postcss": "^8.4.31",
 		"postcss-loader": "^7.3.3",
-		"prettier": "^3.0.2",
+		"prettier": "^3.0.3",
 		"react-refresh": "^0.14.0",
 		"react-test-renderer": "^18.2.0",
-		"rimraf": "^5.0.1",
-		"sass": "^1.66.1",
+		"rimraf": "^5.0.5",
+		"sass": "^1.68.0",
 		"sass-loader": "^13.3.2",
 		"style-loader": "^3.3.3",
 		"tailwindcss": "^3.3.3",
@@ -122,10 +122,10 @@
 		"ts-loader": "^9.4.4",
 		"ts-node": "^10.9.1",
 		"tsconfig-paths-webpack-plugin": "^4.1.0",
-		"typescript": "^5.1.6",
+		"typescript": "^5.2.2",
 		"url-loader": "^4.1.1",
 		"webpack": "^5.88.2",
-		"webpack-bundle-analyzer": "^4.9.0",
+		"webpack-bundle-analyzer": "^4.9.1",
 		"webpack-cli": "^5.1.4",
 		"webpack-dev-server": "^4.15.1",
 		"webpack-merge": "^5.9.0"
